refactor: move dndtable srp
