+++ src/components/organisms/DndTable/index.tsx
@@ -2,17 +2,15 @@
 import { HolderOutlined } from '@ant-design/icons';
 import type { DragEndEvent } from '@dnd-kit/core';
 import { DndContext } from '@dnd-kit/core';
-import type { SyntheticListenerMap } from '@dnd-kit/core/dist/hooks/utilities';
 import { restrictToVerticalAxis } from '@dnd-kit/modifiers';
 import {
   SortableContext,
-  useSortable,
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
 import { Button, TablePaginationConfig } from 'antd';
-import { createContext, useContext, useMemo } from 'react';
-import { FHTable } from './Table';
+import { useContext } from 'react';
+import { DndTableRow, RowContext } from './DndTableRow';
+import { FHTable } from '@finhub/components/atoms/Table';
 
 export interface TableProps {
   dataSource: any;
@@ -25,13 +23,6 @@ export interface TableProps {
   restrictToVerticalAxis?: any;
   onDragEnd?: ({ active, over }: DragEndEvent) => void;
 }
-
-interface RowContextProps {
-  setActivatorNodeRef?: (element: HTMLElement | null) => void;
-  listeners?: SyntheticListenerMap;
-}
-
-const RowContext = createContext<RowContextProps>({});
 
 const DragHandle: React.FC = () => {
   const { setActivatorNodeRef, listeners } = useContext(RowContext);
@@ -47,40 +38,6 @@ const DragHandle: React.FC = () => {
   );
 };
 
-interface RowProps extends React.HTMLAttributes<HTMLTableRowElement> {
-  'data-row-key': string;
-}
-
-const Row: React.FC<RowProps> = (props: RowProps) => {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    setActivatorNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id: props['data-row-key'] });
-
-  const style: React.CSSProperties = {
-    ...props.style,
-    transform: CSS.Translate.toString(transform),
-    transition,
-    ...(isDragging ? { position: 'relative', zIndex: 9999 } : {}),
-  };
-
-  const contextValue = useMemo<RowContextProps>(
-    () => ({ setActivatorNodeRef, listeners }),
-    [setActivatorNodeRef, listeners],
-  );
-
-  return (
-    <RowContext.Provider value={contextValue}>
-      <tr {...props} ref={setNodeRef} style={style} {...attributes} />
-    </RowContext.Provider>
-  );
-};
-
 export const FHDndTable = ({
   dataSource,
   columns,
@@ -91,7 +48,6 @@ export const FHDndTable = ({
   defaultPageSize,
   onDragEnd,
 }: TableProps) => {
-  // const [data, setData] = useState(dataSource);
   const handleDragEnd = (event: DragEndEvent) => {
     if (onDragEnd) {
       onDragEnd(event);
@@ -122,7 +78,7 @@ export const FHDndTable = ({
           defaultPageSize={defaultPageSize}
           onRow={onRow}
           rowKey="key"
-          components={{ body: { row: Row } }}
+          components={{ body: { row: DndTableRow } }}
         />
       </SortableContext>
     </DndContext>
